backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.30
pymysql==1.1.1
python-dotenv==1.0.1
pydantic[email]==2.7.1
apscheduler==3.10.4
requests==2.32.3
python-dateutil==2.9.0.post0
joblib==1.4.2
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.0
xgboost==2.0.3
imbalanced-learn==0.12.3
